Refactor: move quorum into mod membership
