+++ openraft/src/membership/mod.rs
@@ -4,4 +4,6 @@ mod membership;
 #[cfg(test)]
 mod membership_test;
 
+pub mod quorum;
+
 pub use membership::Membership;
